requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
click==8.2.0
crispy-bootstrap5==2025.4
Django==5.2
django-crispy-forms==2.4
djangorestframework==3.16.0
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
